Refactor recursive CTE execution state
