+++ src/execution/operator/set/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_library_unity(
   duckdb_operator_set OBJECT physical_union.cpp physical_recursive_cte.cpp
-  physical_recursive_cte_runtime.cpp physical_cte.cpp)
+  physical_recursive_cte_runtime.cpp physical_recursive_cte_state.cpp
+  physical_cte.cpp)
 set(ALL_OBJECT_FILES
     ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:duckdb_operator_set>
     PARENT_SCOPE)
